Correctly report cellsynt ref id when sending sms
The cellsynt reference ID should have been reported in the "id"
key of the returned hash ref from send_sms(). Instead, the id was
available via the "uri" key. This change sets both keys to the same
value for backwards compatibility. In a future version, "uri" will
be removed.

This issue was reported by plundra.